[WiP} Bug 7826: Data validation failed for path
Modified deleteGroupInternal method by adding isPresent for group-id
to fix the CSIT error of group-id not existing.

+++ mdsalutil/mdsalutil-impl/src/main/java/org/opendaylight/genius/mdsalutil/internal/MDSALManager.java
@@ -272,8 +272,11 @@ public class MDSALManager extends AbstractLifecycle implements IMdsalApiManager 
         Node nodeDpn = buildDpnNode(dpId);
         long groupId = group.getGroupId().getValue();
         InstanceIdentifier<Group> groupInstanceId = buildGroupInstanceIdentifier(groupId, nodeDpn);
-
-        tx.delete(LogicalDatastoreType.CONFIGURATION, groupInstanceId);
+        Optional<Group>existingGroupId = MDSALUtil.read(dataBroker, LogicalDatastoreType.CONFIGURATION,
+                groupInstanceId);
+        if (existingGroupId.isPresent()) {
+            tx.delete(LogicalDatastoreType.CONFIGURATION, groupInstanceId);
+        }
     }
 
     public CheckedFuture<Void, TransactionCommitFailedException> removeFlowInternal(FlowEntity flowEntity) {
